Minor improvements

+++ src/melo_benchmark/evaluation/semantic_baseline/openai_biencoder.py
@@ -48,7 +48,7 @@ class OpenAiBiEncoderScorer(BiEncoderScorer):
     def _compute_embeddings(
                 self,
                 rendered_prompts: List[str]
-            ) -> List[List[int]]:
+            ) -> List[List[float]]:
 
         results = []
         for i in range(0, len(rendered_prompts), self.batch_size):
