+++ src/melo_benchmark/evaluation/semantic_baseline/stransf_biencoder.py
@@ -31,7 +31,7 @@ class SentenceTransformersBiEncoderScorer(BiEncoderScorer):
     def _compute_embeddings(
                 self,
                 rendered_prompts: List[str]
-            ) -> List[List[int]]:
+            ) -> List[List[float]]:
 
         results = []
         for i in range(0, len(rendered_prompts), self.batch_size):
