+++ src/melo_benchmark/evaluation/semantic_baseline/tfhub_biencoder.py
@@ -45,7 +45,7 @@ class TFHubBiEncoderScorer(BiEncoderScorer):
     def _compute_embeddings(
                 self,
                 rendered_prompts: List[str]
-            ) -> List[List[int]]:
+            ) -> List[List[float]]:
 
         results = []
         for i in range(0, len(rendered_prompts), self.batch_size):
